# src/main/java/wasserballturnier/api/generated/api/MannschaftApiController.java
package wasserballturnier.api.generated.api;

import com.fasterxml.jackson.databind.ObjectMapper;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.bind.annotation.CookieValue;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.multipart.MultipartFile;
import wasserballturnier.api.generated.model.MannschaftArrayTO;
import wasserballturnier.api.generated.model.MannschaftPushTO;
import wasserballturnier.api.persistence.mannschaft.Mannschaft;
import wasserballturnier.api.services.MannschaftService;

import javax.validation.constraints.*;
import javax.validation.Valid;
import javax.servlet.http.HttpServletRequest;
import java.io.IOException;
import java.util.List;
import java.util.Map;

@RestController
public class MannschaftApiController implements MannschaftApi {

    private static final Logger log = LoggerFactory.getLogger(MannschaftApiController.class);

    private final MannschaftService mannschaftService;

    public MannschaftApiController(MannschaftService mannschaftService) {
        this.mannschaftService = mannschaftService;
    }

    public ResponseEntity<Integer> mannschaftPut(@Parameter(in = ParameterIn.DEFAULT, description = "", required=true, schema=@Schema()) @Valid @RequestBody MannschaftPushTO body) {
        return ResponseEntity.ok(mannschaftService.addMannschaft(new Mannschaft(body.getMannschaftsname())));
    }

    public ResponseEntity<MannschaftArrayTO> mannschaftenPost() {
        return new ResponseEntity<MannschaftArrayTO>(HttpStatus.NOT_IMPLEMENTED);
    }

}